notification_service/requirements.txt
Flask===3.0.0
pika===1.3.2
requests===2.31.0
pymongo[srv]===4.5.0
websockets==12.0
flask-cors==3.0.10
PyJWT===2.8.0
aio-pika==9.4.2
motor==3.0.0

